requirements.txt
vosk==0.3.45
pyttsx3==2.90
openai==1.3.0
langchain==0.0.350
sounddevice==0.4.6
numpy==1.24.3
pandas==2.0.3
python-dotenv==1.0.0
gspread==5.12.0
oauth2client==4.1.3 